refactor: modularize schemas into schemas/runtime and schemas/graph
Centralize schema definitions for better reusability and separation of concerns:

Schema structure:
- schemas/runtime/ - decision.py, run.py (Decision, Run, RunStatus, etc.)
- schemas/graph/ - goal.py, node.py, edge.py, plan.py (Goal, Plan, GraphSpec, etc.)

Key changes:
- Move decision/run schemas from schemas/ to schemas/runtime/
- Create schemas/graph/ with goal, node, edge, plan data models
- graph/edge.py now subclasses EdgeSpec/GraphSpec and adds behavior methods
- graph/goal.py, node.py, plan.py re-export from schemas for backward compat
- schemas/__init__.py re-exports all types as single entry point
- Update imports across runtime, storage, builder, and tests

All 438 tests passing.

Fixes #1568
