+++ core/framework/__init__.py
@@ -26,8 +26,15 @@ from framework.builder.query import BuilderQuery
 from framework.llm import AnthropicProvider, LLMProvider
 from framework.runner import AgentOrchestrator, AgentRunner
 from framework.runtime.core import Runtime
-from framework.schemas.decision import Decision, DecisionEvaluation, Option, Outcome
-from framework.schemas.run import Problem, Run, RunSummary
+from framework.schemas import (
+    Decision,
+    DecisionEvaluation,
+    Option,
+    Outcome,
+    Problem,
+    Run,
+    RunSummary,
+)
 
 # Testing framework
 from framework.testing import (
