+++ core/framework/schemas/__init__.py
@@ -1,14 +1,86 @@
-"""Schema definitions for runtime data."""
+"""
+Schema definitions - single entry point for framework schemas.
 
-from framework.schemas.decision import Decision, DecisionEvaluation, Option, Outcome
-from framework.schemas.run import Problem, Run, RunSummary
+Subpackages:
+- schemas.runtime  → decision.py, run.py (Decision, Run, RunStatus, etc.)
+- schemas.graph   → goal.py, node.py, edge.py, plan.py (Goal, Plan, GraphSpec, etc.)
+
+Import from here for "all schemas", or from framework.schemas.runtime / framework.schemas.graph
+for a specific domain.
+"""
+
+from framework.schemas.graph import (
+    ActionSpec,
+    ActionType,
+    ApprovalDecision,
+    ApprovalRequest,
+    ApprovalResult,
+    AsyncEntryPointSpec,
+    Constraint,
+    EdgeCondition,
+    EdgeSpec,
+    EvaluationRule,
+    ExecutionStatus,
+    Goal,
+    GoalStatus,
+    GraphSpec,
+    Judgment,
+    JudgmentAction,
+    NodeSpec,
+    Plan,
+    PlanExecutionResult,
+    PlanStep,
+    StepStatus,
+    SuccessCriterion,
+)
+from framework.schemas.runtime.decision import (
+    Decision,
+    DecisionEvaluation,
+    DecisionType,
+    Option,
+    Outcome,
+)
+from framework.schemas.runtime.run import (
+    Problem,
+    Run,
+    RunMetrics,
+    RunStatus,
+    RunSummary,
+)
 
 __all__ = [
+    # Runtime
     "Decision",
+    "DecisionEvaluation",
+    "DecisionType",
     "Option",
     "Outcome",
-    "DecisionEvaluation",
-    "Run",
-    "RunSummary",
     "Problem",
+    "Run",
+    "RunMetrics",
+    "RunStatus",
+    "RunSummary",
+    # Graph
+    "ActionSpec",
+    "ActionType",
+    "ApprovalDecision",
+    "ApprovalRequest",
+    "ApprovalResult",
+    "AsyncEntryPointSpec",
+    "Constraint",
+    "EdgeCondition",
+    "EdgeSpec",
+    "EvaluationRule",
+    "ExecutionStatus",
+    "Goal",
+    "GoalStatus",
+    "GraphSpec",
+    "Judgment",
+    "JudgmentAction",
+    "NodeSpec",
+    "Plan",
+    "PlanExecutionResult",
+    "PlanStep",
+    "StepStatus",
+    "SuccessCriterion",
 ]
