+++ core/framework/schemas/graph/__init__.py
@@ -1,0 +1,50 @@
+"""Graph schemas - goal, node, edge, plan."""
+
+from framework.schemas.graph.edge import (
+    AsyncEntryPointSpec,
+    EdgeCondition,
+    EdgeSpec,
+    GraphSpec,
+)
+from framework.schemas.graph.goal import Constraint, Goal, GoalStatus, SuccessCriterion
+from framework.schemas.graph.node import NodeSpec
+from framework.schemas.graph.plan import (
+    ActionSpec,
+    ActionType,
+    ApprovalDecision,
+    ApprovalRequest,
+    ApprovalResult,
+    EvaluationRule,
+    ExecutionStatus,
+    Judgment,
+    JudgmentAction,
+    Plan,
+    PlanExecutionResult,
+    PlanStep,
+    StepStatus,
+)
+
+__all__ = [
+    "Constraint",
+    "Goal",
+    "GoalStatus",
+    "SuccessCriterion",
+    "NodeSpec",
+    "AsyncEntryPointSpec",
+    "EdgeCondition",
+    "EdgeSpec",
+    "GraphSpec",
+    "ActionSpec",
+    "ActionType",
+    "ApprovalDecision",
+    "ApprovalRequest",
+    "ApprovalResult",
+    "EvaluationRule",
+    "ExecutionStatus",
+    "Judgment",
+    "JudgmentAction",
+    "Plan",
+    "PlanExecutionResult",
+    "PlanStep",
+    "StepStatus",
+]
