+++ core/tests/test_concurrent_storage.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from framework.schemas.run import Run, RunMetrics, RunStatus
+from framework.schemas import Run, RunMetrics, RunStatus
 from framework.storage.concurrent import ConcurrentStorage
